pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "EmotionRecognition"
dynamic = ["version"]
description = "application"
readme = "README.md"
requires-python = ">=3.8"
keywords = ['tkinter', 'app', 'GUI', ]
authors = [
    { name = "shalabycr7", email = "abdoshalaby.dev@gmail.com" },
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: Implementation :: CPython",
    "Programming Language :: Python :: Implementation :: PyPy",
]
dependencies = [
    'ttkbootstrap~=1.10.1',
    'opencv-python~=4.9.0.80',
    'numpy~=1.26.4',
    'pillow~=10.3.0',
    'keras~=3.3.3',
    'setuptools~=69.5.1',
    'tensorflow~=2.16.1',
    'pandas~=2.2.2',
    'matplotlib~=3.9.0',
    'seaborn~=0.13.2',
    'scikit-learn~=1.5.0',
    'future~=1.0.0',
]

[project.urls]
Documentation = "https://github.com/unknown/EmotionRecognition#readme"
Issues = "https://github.com/unknown/EmotionRecognition/issues"
Source = "https://github.com/unknown/EmotionRecognition"

[tool.hatch.version]
path = "src/EmotionRecognition/__about__.py"

[tool.hatch.envs.default]
dependencies = [
    "coverage[toml]>=6.5",
    "pytest",
]
[tool.hatch.envs.default.scripts]
test = "pytest {args:tests}"
test-cov = "coverage run -m pytest {args:tests}"
cov-report = [
    "- coverage combine",
    "coverage report",
]
cov = [
    "test-cov",
    "cov-report",
]

[[tool.hatch.envs.all.matrix]]
python = ["3.8", "3.9", "3.10", "3.11", "3.12"]

[tool.hatch.envs.types]
dependencies = [
    "mypy>=1.0.0",
]
[tool.hatch.envs.types.scripts]
check = "mypy --install-types --non-interactive {args:src/EmotionRecognition tests}"

[tool.coverage.run]
source_pkgs = ["EmotionRecognition", "tests"]
branch = true
parallel = true
omit = [
    "src/EmotionRecognition/__about__.py",
]

[tool.coverage.paths]
EmotionRecognition = ["src/EmotionRecognition", "*/EmotionRecognition/src/EmotionRecognition"]

[tool.coverage.report]
exclude_lines = [
    "no cov",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
]